+++ src/main/java/RangerCaptain/cardfusion/components/WeakComponent.java
@@ -13,11 +13,11 @@ public class WeakComponent extends AbstractSimpleApplyComponent {
     public static final String[] DESCRIPTION_TEXT = CardCrawlGame.languagePack.getUIString(ID).TEXT;
     public static final String[] CARD_TEXT = CardCrawlGame.languagePack.getUIString(ID).EXTRA_TEXT;
 
-    public WeakComponent(int base) {
+    public WeakComponent(float base) {
         this(base, ComponentTarget.ENEMY);
     }
 
-    public WeakComponent(int base, ComponentTarget target) {
+    public WeakComponent(float base, ComponentTarget target) {
         super(ID, base, target);
     }
